refactor: Expect conversation ID when creating messages [BREAKING]
